% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-wcm.R
\name{print.wcmtools_palette}
\alias{print.wcmtools_palette}
\title{Print a WCM palette}
\usage{
\method{print}{wcmtools_palette}(x, ..., family = getOption("wcmtools.font_family", "WCM Sans"))
}
\arguments{
\item{x}{A \code{"wcmtools_palette"} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints a color swatch for objects returned by \code{\link[=palette_wcm]{palette_wcm()}}.
}
