% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{wcm_disable_showtext}
\alias{wcm_disable_showtext}
\title{Disable showtext for plotting (revert to default device fonts)}
\usage{
wcm_disable_showtext()
}
\description{
Disable showtext for plotting (revert to default device fonts)
}
