% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-preamble.R
\name{wcm_setup}
\alias{wcm_setup}
\title{Set knitr defaults/fonts for WCM reports}
\usage{
wcm_setup(attach_html_font = FALSE, register_plots = TRUE, ...)
}
\arguments{
\item{attach_html_font}{Logical; call \code{\link[=wcm_attach_font]{wcm_attach_font()}} for HTML font dependencies.}

\item{register_plots}{Logical; call \code{\link[=wcm_register_fonts]{wcm_register_fonts()}} so plots use WCM Sans.}

\item{...}{Optional overrides for \code{knitr::opts_chunk$set()}.}
}
\value{
List of set knitr options.
}
\description{
Override any knitr option via \code{...}, or set global defaults:
\code{options(wcmtools.knitr_defaults = list(fig.width = 8))}
}
