% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{wcm_with_showtext}
\alias{wcm_with_showtext}
\title{Wrapper to temporarily enable showtext (for \code{expr} only)}
\usage{
wcm_with_showtext(expr)
}
\description{
Wrapper to temporarily enable showtext (for \code{expr} only)
}
